refactor: move code of sorting and border tasks
