+++ src/main.js
@@ -1,6 +1,8 @@
 import {renderMenuComponent} from './view/menu.js';
 import {renderFiltersComponent} from './view/filters.js';
 import {renderBoardComponent} from './view/board.js';
+import {renderSortingComponent} from './view/sorting.js';
+import {renderBoardTasksComponent} from './view/board-tasks.js';
 
 const COUNT_OF_TASKS = 3;
 
@@ -10,22 +12,6 @@ const renderComponent = (container, template, place = `beforeend`) => {
 
 const siteMainElement = document.querySelector(`.main`);
 const siteMainControl = siteMainElement.querySelector(`.main__control`);
-
-const renderSortingComponent = () => {
-  return `
-    <div class="board__filter-list">
-      <a href="#" class="board__filter" data-sort-type="default">SORT BY DEFAULT</a>
-      <a href="#" class="board__filter" data-sort-type="date-up">SORT BY DATE up</a>
-      <a href="#" class="board__filter" data-sort-type="date-down">SORT BY DATE down</a>
-    </div>
-  `;
-};
-
-const renderBoardTasksComponent = () => {
-  return `
-    <div class="board__tasks"></div>
-  `;
-};
 
 const renderEditTaskComponent = () => {
   return `
