+++ src/view/board-tasks.js
@@ -1,0 +1,7 @@
+export {renderBoardTasksComponent};
+
+const renderBoardTasksComponent = () => {
+  return `
+    <div class="board__tasks"></div>
+  `;
+};
